+++ cassandra/src/main/java/com/yahoo/ycsb/db/CassandraCQLClient.java
@@ -260,7 +260,7 @@ public class CassandraCQLClient extends DB {
    * @return Zero on success, a non-zero error code on error
    */
   @Override
-  public Status scan(String table, String startkey, int recordcount,
+  public Status scan(String table, String startkey, long recordcount,
       Set<String> fields, Vector<HashMap<String, ByteIterator>> result) {
 
     try {
