+++ core/src/main/java/com/yahoo/ycsb/Client.java
@@ -366,7 +366,7 @@ class ClientThread implements Runnable
   DB _db;
   boolean _dotransactions;
   Workload _workload;
-  int _opcount;
+  long _opcount;
   double _targetOpsPerMs;
 
   int _opsdone;
@@ -388,7 +388,7 @@ class ClientThread implements Runnable
    * @param targetperthreadperms target number of operations per thread per ms
    * @param completeLatch The latch tracking the completion of all clients.
    */
-  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, int opcount, double targetperthreadperms, CountDownLatch completeLatch)
+  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, long opcount, double targetperthreadperms, CountDownLatch completeLatch)
   {
     _db=db;
     _dotransactions=dotransactions;
@@ -529,9 +529,9 @@ class ClientThread implements Runnable
   /**
    * the total amount of work this thread is still expected to do
    */
-  public int getOpsTodo()
+  public long getOpsTodo()
   {
-    int todo = _opcount - _opsdone;
+    long todo = _opcount - _opsdone;
     return todo < 0 ? 0 : todo;
   }
 }
@@ -663,7 +663,7 @@ public class Client
    * loaded from conf.
    * @throws IOException Either failed to write to output stream or failed to close it.
    */
-  private static void exportMeasurements(Properties props, int opcount, long runtime)
+  private static void exportMeasurements(Properties props, long opcount, long runtime)
     throws IOException
   {
     MeasurementsExporter exporter = null;
@@ -1033,20 +1033,20 @@ public class Client
     boolean initFailed = false;
     try (final TraceScope span = tracer.newScope(CLIENT_INIT_SPAN)) {
 
-      int opcount;
+      long opcount;
       if (dotransactions)
       {
-        opcount=Integer.parseInt(props.getProperty(OPERATION_COUNT_PROPERTY,"0"));
+        opcount=Long.parseLong(props.getProperty(OPERATION_COUNT_PROPERTY,"0"));
       }
       else
       {
         if (props.containsKey(INSERT_COUNT_PROPERTY))
         {
-          opcount=Integer.parseInt(props.getProperty(INSERT_COUNT_PROPERTY,"0"));
+          opcount=Long.parseLong(props.getProperty(INSERT_COUNT_PROPERTY,"0"));
         }
         else
         {
-          opcount=Integer.parseInt(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
+          opcount=Long.parseLong(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
         }
       }
 
@@ -1065,7 +1065,7 @@ public class Client
         }
 
 
-        int threadopcount = opcount / threadcount;
+        long threadopcount = opcount / threadcount;
 
         // ensure correct number of operations, in case opcount is not a multiple of threadcount
         if (threadid<opcount%threadcount)
