+++ infinispan/src/main/java/com/yahoo/ycsb/db/InfinispanRemoteClient.java
@@ -93,7 +93,7 @@ public class InfinispanRemoteClient extends DB {
   }
 
   @Override
-  public Status scan(String table, String startkey, int recordcount, Set<String> fields,
+  public Status scan(String table, String startkey, long recordcount, Set<String> fields,
                      Vector<HashMap<String, ByteIterator>> result) {
     LOGGER.warn("Infinispan does not support scan semantics");
     return Status.NOT_IMPLEMENTED;
