+++ orientdb/src/main/java/com/yahoo/ycsb/db/OrientDBClient.java
@@ -268,7 +268,7 @@ public class OrientDBClient extends DB {
   }
 
   @Override
-  public Status scan(String table, String startkey, int recordcount, Set<String> fields,
+  public Status scan(String table, String startkey, long recordcount, Set<String> fields,
                      Vector<HashMap<String, ByteIterator>> result) {
     if (isRemote) {
       // Iterator methods needed for scanning are Unsupported for remote database connections.
